requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
django-etc==1.3.1
django-heroku==0.3.1
django-hitcount==1.3.5
Pillow==9.1.1
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
wcwidth==0.2.5
webencodings==0.5.1
websockets==3.2
whitenoise==6.2.0
